Bug 521788 - spawner doesnt kill child process on MacOS X
Fixed by copying code from org.eclipse.cdt.core.linux that
contains the fix from bug 119387 : set the process group by
calling setpgid() when forking a process, and use killpg() for
killing the entire process group on termination.

This also aligns MacOS code fully with Linux code. Tested on
MacOSX 10.8 and 10.13 beta. Binaries will be submitted with the
related fix for bug 519886.

+++ core/org.eclipse.cdt.core.macosx/library/spawner.c
@@ -7,6 +7,8 @@
  *
  * Contributors:
  *     QNX Software Systems - initial API and implementation
+ *     Wind River Systems, Inc.  
+ *     Mikhail Zabaluev (Nokia) - bug 82744
  *******************************************************************************/
 #include <unistd.h>
 #include <stdlib.h>
@@ -90,52 +92,53 @@ JNIEXPORT jint JNICALL Java_org_eclipse_cdt_utils_spawner_Spawner_exec2
   (JNIEnv *env, jobject jobj, jobjectArray jcmd, jobjectArray jenv, jstring jdir, jintArray jchannels,
    jstring jslaveName, jint masterFD, jboolean console)
 {
-	jint *channels = (*env)->GetIntArrayElements(env, jchannels, 0);
-	const char *dirpath = (*env)->GetStringUTFChars(env, jdir, NULL);
-	const char *pts_name = (*env)->GetStringUTFChars(env, jslaveName, NULL);
-	char **cmd = NULL;
-	char **envp = NULL;
-	int fd[3];
-	pid_t pid = -1;
-                                                                                                                             
-	if (channels == NULL)
-		goto bail_out;
-                                                                                                                             
-	cmd = alloc_c_array(env, jcmd);
-	if (cmd == NULL)
-		goto bail_out;
- 
-	envp = alloc_c_array(env, jenv);
-	if (envp == NULL)
-		goto bail_out;
- 
+    jint *channels = (*env)->GetIntArrayElements(env, jchannels, 0);
+    const char *dirpath = (*env)->GetStringUTFChars(env, jdir, NULL);
+    const char *pts_name = (*env)->GetStringUTFChars(env, jslaveName, NULL);
+    char **cmd = NULL;
+    char **envp = NULL;
+    int fd[3];
+    pid_t pid = -1;
+
+    if (channels == NULL)
+        goto bail_out;
+
+    cmd = alloc_c_array(env, jcmd);
+    if (cmd == NULL)
+        goto bail_out;
+
+    envp = alloc_c_array(env, jenv);
+    if (envp == NULL)
+        goto bail_out;
+
 #if DEBUGIT
-	fprintf(stderr, "command:");
-	print_array(cmd);
-	fprintf(stderr, "Envp:");
-	print_array(envp);
-	fprintf(stderr, "dirpath: %s\n", dirpath);
-	fprintf(stderr, "pts_name: %s\n", pts_name);
+    fprintf(stderr, "command:");
+    print_array(cmd);
+    fprintf(stderr, "Envp:");
+    print_array(envp);
+    fprintf(stderr, "dirpath: %s\n", dirpath);
+    fprintf(stderr, "pts_name: %s\n", pts_name);
 #endif
- 
-	pid = exec_pty(cmd[0], cmd, envp, dirpath, fd, pts_name, masterFD, console);
-	if (pid < 0)
-		goto bail_out;
- 
-	channels[0] = fd[0];
-	channels[1] = fd[1];
-	channels[2] = fd[2];
- 
- bail_out:
-	(*env)->ReleaseIntArrayElements(env, jchannels, channels, 0);
-	(*env)->ReleaseStringUTFChars(env, jdir, dirpath);
-	(*env)->ReleaseStringUTFChars(env, jslaveName, pts_name);
-	if (cmd)
-		free_c_array(cmd);
-	if (envp)
-		free_c_array(envp);
-	return pid;
+
+    pid = exec_pty(cmd[0], cmd, envp, dirpath, fd, pts_name, masterFD, console);
+    if (pid < 0)
+        goto bail_out;
+
+    channels[0] = fd[0];
+    channels[1] = fd[1];
+    channels[2] = fd[2];
+
+  bail_out:
+    (*env)->ReleaseIntArrayElements(env, jchannels, channels, 0);
+    (*env)->ReleaseStringUTFChars(env, jdir, dirpath);
+    (*env)->ReleaseStringUTFChars(env, jslaveName, pts_name);
+    if (cmd)
+        free_c_array(cmd);
+    if (envp)
+        free_c_array(envp);
+    return pid;
 }
+ 
 
 JNIEXPORT jint JNICALL
 Java_org_eclipse_cdt_utils_spawner_Spawner_exec1(JNIEnv * env, jobject jobj,
@@ -246,23 +249,38 @@ Java_org_eclipse_cdt_utils_spawner_Spawner_raise(JNIEnv * env, jobject jobj,
 
     switch (sig) {
     case 0:                    /* NOOP */
-        status = kill(pid, 0);
+        status = killpg(pid, 0);
+        if(status == -1) {
+	        status = kill(pid, 0);
+        }
 	break;
 
     case 2:                    /* INTERRUPT */
-        status = kill(pid, SIGINT);
+        status = killpg(pid, SIGINT);
+        if(status == -1) {
+	        status = kill(pid, SIGINT);
+        }
 	break;
 
     case 9:                    /* KILL */
-        status = kill(pid, SIGKILL);
+        status = killpg(pid, SIGKILL);
+        if(status == -1) {
+        	status = kill(pid, SIGKILL);
+        }
 	break;
 
     case 15:                   /* TERM */
-        status = kill(pid, SIGTERM);
+        status = killpg(pid, SIGTERM);
+        if(status == -1) {
+    	    status = kill(pid, SIGTERM);
+        }
 	break;
 
     default:
-        status = kill(pid, sig);        /* WHAT ?? */
+        status = killpg(pid, sig);        /* WHAT ?? */
+        if(status == -1) {
+	        status = kill(pid, sig);        /* WHAT ?? */
+        }
 	break;
     }
